+++ CHANGELOG.md
@@ -3,7 +3,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 1.18.1-1.0.0
+
+### Added
+- Whitelist of items in the uncrafter (if the list contains value the blacklist isn't used).
+- Italian translation (thanks to SimGitHub5)
+
+### Changed
+- Blacklist now uses items instead of recipes (whitelist uses items too)
+- InterModCommunication now uses items instead of recipes
+
+### Fixed
+- Uncrafting items with tags as ingredient give random items
+
 ## 1.18-0.4.0
 
 - ported version 1.17.1-0.4.0 to 1.18
--